+++ src/components/CastContainer.tsx
@@ -1,0 +1,30 @@
+import { useState, useEffect } from 'react';
+import { APIKEY } from '../Services/api-client';
+import { Credits } from '../types/CastTypes';
+import CastPresentational from './CastPresentational';
+
+interface CastProps {
+    movie_id: number;
+    page: number;
+}
+
+const CastContainer = ({ movie_id, page }: CastProps) => {
+    const [credits, setCredits] = useState<Credits | null>(null);
+    
+    useEffect(() => {
+        fetch(
+            `https://api.themoviedb.org/3/movie/${movie_id}/credits?api_key=${APIKEY}&page=${page}`
+        )
+            .then((response) => response.json())
+            .then((data) => setCredits(data))
+            .catch((error) => console.log(error));
+    }, [movie_id, page]);
+
+    return (
+        <>
+            <CastPresentational credits={credits}/>
+        </>
+    );
+};
+
+export default CastContainer;
