+++ src/components/ModelCard.tsx
@@ -1,6 +1,6 @@
 
 import { GoArrowLeft } from "react-icons/go"
-import CastMt from "./CastMT"
+import CastMt from "./CastContainer"
 import { img_500, unavailable } from "../Services/Config"
 import Trailer from "./Trailer"
 import { ModalProps } from "../types/ModalTypes"
