+++ src/components/SliderPresentational.tsx
@@ -1,0 +1,11 @@
+
+
+const SliderPresentational = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default SliderPresentational
