+++ src/index.css
@@ -605,7 +605,7 @@ a:hover {
 .cast-actor-grid img {
     background-image: url('https://user-images.githubusercontent.com/10515204/56117400-9a911800-5f85-11e9-878b-3f998609a6c8.jpg');
     width: 100%;
-    height: 100%;
+    height: 280px;
     background-repeat: no-repeat;
     object-fit: cover;
 }
